First commit in a long time. Next step, writing tests and refactoring

+++ BugTracker/Models/Services/Bugs/BugService.cs
@@ -2,60 +2,91 @@
 using BugTracker.Models.Bugs.Severity;
 using BugTracker.Models.Bugs.Status;
 using BugTracker.Models.Repositories;
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
-using System;
-using BugTracker.Models.Repositories.Bugs;
 
 namespace BugTracker.Models.Services.Bugs
 {
     public class BugService : IBugService
     {
-        public IBugRepository Repository { get; }
+        public IRepository<Bug> Repository { get; }
 
-        public BugService(IBugRepository repository)
+        public BugService(IRepository<Bug> repository)
         {
             Repository = repository;
         }
-        public async Task<IEnumerable<Bug>> GetBugsByPriority(IBugSeverity priorty)
+
+
+        public async Task<IEnumerable<Bug>> GetBugsByPriorityAsync(BugSeverity priorty, bool asTrackable = true)
         {
-            var allBugs = await Repository.GetAll();
-            return allBugs.Where(bug => bug.Severity == priorty);
+            if (priorty == null) throw new ArgumentNullException();
+            Expression<Func<Bug, bool>> expression = bug => bug.Severity == priorty;
+            return await Repository.GetByQueryAsync(expression, asTrackable);
         }
 
-        public async Task<IEnumerable<Bug>> GetBugsByStatus(IBugStatus status)
+        public IEnumerable<Bug> GetBugsByPriority(BugSeverity priorty, bool asTrackable = true)
         {
-            var allBugs = await Repository.GetAll();
-            return allBugs.Where(bug => bug.Status == status);
+            throw new NotImplementedException();
         }
 
-        public async Task<IEnumerable<Bug>> GetAllBugs()
-        => await Repository.GetAll();
-
-        public async Task<bool> DeleteBug(Bug bug)
+        public async Task<IEnumerable<Bug>> GetBugsByStatusAsync(BugStatus status, bool asTrackable = true)
         {
-            var isDeleted =  Repository.Delete(bug);
-            if (isDeleted)
-                await Repository.SaveChangesAsync();
-            return isDeleted;
+            if (status == null) throw new ArgumentNullException();
+            Expression<Func<Bug, bool>> expression = bug => bug.Status == status;
+            return await Repository.GetByQueryAsync(expression, asTrackable);
         }
 
-        public async Task<bool> UpdateBug(Bug bug)
+
+        public IEnumerable<Bug> GetBugsByStatus(BugStatus status, bool asTrackable = true)
         {
-            var isModified =  Repository.Update(bug);
-            if (isModified)
-                await Repository.SaveChangesAsync();
-            return isModified;
+            throw new NotImplementedException();
         }
 
-        public async Task<int> AddBug(Bug bug)
+        public async Task<IEnumerable<Bug>> GetAllBugsAsync(bool asTrackable = true)
+          => await Repository.GetAllAsync(asTrackable);
+
+        public IEnumerable<Bug> GetAllBugs(bool asTrackable)
+          => Repository.GetAll(asTrackable);
+
+        public async Task DeleteBugAsync(Bug bug)
         {
+            if (bug == null) throw new BugNotFoundException();
+            Repository.Delete(bug);
+            await Repository.SaveChangesAsync();
+        }
+
+        public void DeleteBug(Bug bug)
+        {
+            if (bug == null) throw new BugNotFoundException();
+            Repository.Delete(bug);
+            Repository.SaveChanges();
+        }
+
+        public async Task UpdateBugAsync(Bug bug)
+        {
+            if (bug == null) throw new BugNotFoundException();
+            Repository.Update(bug);
+            await Repository.SaveChangesAsync();
+        }
+
+        public void UpdateBug(Bug bug)
+        {
+            if (bug == null) throw new BugNotFoundException();
+            Repository.Update(bug);
+            Repository.SaveChanges();
+        }
+
+        public async Task<int> AddBugAsync(Bug bug)
+        {
+            if (bug == null) throw new BugNotFoundException();
             try
             {
-                var bugId = Repository.Add(bug);
+                Repository.Add(bug);
                 await Repository.SaveChangesAsync();
-                return bugId;
+                return bug.Id;
             }
             catch (ArgumentNullException)
             {
@@ -63,15 +94,51 @@ namespace BugTracker.Models.Services.Bugs
             }
         }
 
-        public async Task<Bug> GetBugById(int? id)
+        public int AddBug(Bug bug)
+        {
+            if (bug == null) throw new BugNotFoundException();
+            try
+            {
+                Repository.Add(bug);
+                Repository.SaveChanges();
+                return bug.Id;
+            }
+            catch (ArgumentNullException)
+            {
+                throw new BugNotAddedException();
+            }
+        }
+
+        public async Task<Bug> GetBugByIdAsync(int? id, bool asTrackable = true)
         {
             if (id == null) throw new BugNotFoundException();
-            Bug bugFound = await Repository.GetById(id.Value);
+            var bugFound = await Repository.GetByIdAsync(id.Value, asTrackable);
             if (bugFound == null) throw new BugNotFoundException();
             return bugFound;
         }
 
-        public async Task<bool> BugExists(int id) 
-            => await Repository.Exists(id);
+        public Bug GetBugById(int? id, bool asTrackable = true)
+        {
+            if (id == null) throw new BugNotFoundException();
+            var bugFound = Repository.GetById(id.Value, asTrackable);
+            if (bugFound == null) throw new BugNotFoundException();
+            return bugFound;
+        }
+
+        public async Task<bool> BugExistsAsync(int id)
+            => await Repository.ExistsAsync(id);
+
+        public bool BugExists(int id) => Repository.Exists(id);
+
+
+
+
+
+
+
+
+
+
+
     }
 }
